+++ pykit.js
@@ -1902,16 +1902,18 @@ pykit.LinkedList = {
 		return node.$tailNode;
 	},
 	contains: function(node) {
-		var next = this.headNode;
-		while (next) {
-			if (node == next) {
-				return true;
-			}
-			else {
-				next = next.$tailNode;
-			}
+		return this._nodeList.indexOf(node) != -1;
+	},
+	indexOf: function(matchNode, beginNode) {
+		var i = 0;
+		var node = beginNode || this.headNode;
+		while (node) {
+			// Apparently 1 == "1" in JS
+			if (node === matchNode)
+				return i;
+			node = node.$tailNode;
+			i++;
 		}
-		return false;
 	},
 	findOne: function(key, value, beginNode) {
 		var node = beginNode || this.headNode;
@@ -2228,7 +2230,7 @@ pykit.UI.list = pykit.defUI({
 		}
 	},
 	setActiveLabel: function(label) {
-		this.setActive("label", label)
+		this.setActive("label", label);
 	},
 	setActive: function(key, value) {
 		this.unselectAll();
@@ -2654,13 +2656,22 @@ pykit.UI.table = pykit.defUI({
 pykit.UI.select = pykit.defUI({
 	__name__: "select",
 	$defaults: {
-		tagClass: "uk-form-select",
+		tagClass: "",
 		htmlTag: "SELECT",
 		flex: false,
 		margin : "",
 		size: "",
 		layout: "",
 		listStyle: ""
+	},
+	select: function(target) {
+		if (pykit.isString(target))
+			target = this.getItem(target);
+		target.$selected = true;
+		this._html.selectedIndex = this.indexOf(target);
+	},
+	unselectAll: function() {
+		// Do nothing
 	},
 	template: function(itemConfig) {
 		return itemConfig.label;
@@ -2669,7 +2680,8 @@ pykit.UI.select = pykit.defUI({
 		parentNode.innerHTML = this.template(config);
 	},
 	_itemHTML: function(config) {
-		return pykit.html.createElement("OPTION", {value: config.value});
+		return pykit.html.createElement("OPTION", {
+			value: config.value, selected: config.selected});
 	}
 }, pykit.UI.list);